requirements.txt
aiogram
python-dotenv
langchain
langchain-community
langchain-core
langchain-huggingface
pinecone
huggingface-hub
fpdf
